Extract dependencies to buildSrc
